#1376 - move plan version date formatting out of shared date utils
