+++ src/components/rangeUsePlanPage/versionsList/VersionsDropdownList.tsx
@@ -11,7 +11,8 @@ import classnames from 'classnames';
 import Status from '../../common/Status';
 import { useUser } from '../../../providers/UserProvider';
 import { PrimaryButton, MuiIcon } from '../../common/';
-import { axios, getAuthHeaderConfig, formatPlanVersionDate } from '../../../utils';
+import { axios, getAuthHeaderConfig } from '../../../utils';
+import { formatPlanVersionDate } from '../../../utils/format/planVersionDate';
 import * as API from '../../../constants/api';
 import AttachmentsList from './AttachmentsList';
 import { ATTACHMENT_TYPE } from '../../../constants/variables';
