+++ src/utils/format/date.ts
@@ -1,17 +1,6 @@
 import moment from 'moment';
-import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
-import timezone from 'dayjs/plugin/timezone';
 import { DATE_FORMAT } from '../../constants/variables';
 import { NOT_PROVIDED } from '../../constants/strings';
-
-dayjs.extend(utc);
-dayjs.extend(timezone);
-
-export const PLAN_TIMEZONE = 'America/Vancouver';
-
-export const formatPlanVersionDate = (date: string | Date | null | undefined): string =>
-  date ? dayjs.utc(date).tz(PLAN_TIMEZONE).format('MMM DD YYYY h:mm a') : '';
 
 /**
  * Present the date time in a more readable way
